pyproject.toml
[project]
name = "takenoko"
version = "0.1.0"
description = "A WanVideo2.1(2)-14B-(A)T2V LoRA trainer"
readme = "README.md"
requires-python = ">=3.10, <3.13"
dependencies = [
  "accelerate==1.6.0",
  "av==14.0.1",
  "bitsandbytes==0.45.4",
  "decord>=0.6.0",
  "diffusers==0.32.1",
  "easydict==1.13",
  "einops==0.7.0",
  "fira>=1.0.2",
  "ftfy==6.3.1",
  "hpsv2==1.2.0",
  "huggingface-hub==0.30.0",
  "lycoris-lora>=3.2.0.post2",
  "matplotlib>=3.10.0",
  "mmcv==1.7.2",
  "moviepy==2.2.1",
  "numpy==1.26.4",
  "nvidia-ml-py>=12.575.51",
  "omegaconf>=2.3.0",
  "opencv-python==4.10.0.84",
  "optimum-quanto>=0.2.7",
  "pillow>=10.2.0",
  "pillow-avif-plugin>=1.5.2",
  "pillow-jxl-plugin>=1.3.4",
  "prodigyopt>=1.1.2",
  "pynvml>=12.0.0",
  "pytorch-optimizer>=3.6.1",
  "requests>=2.31.0",
  "safetensors==0.4.5",
  "scikit-learn>=1.7.1",
  "tensorboard>=2.19.0",
  "tensorboardx>=2.6.4",
  "termcolor>=3.1.0",
  "timm>=1.0.19",
  "toml==0.10.2",
  "torch-optimi>=0.3.0",
  "torch>=2.6.0,<2.7.0",
  "torchmetrics>=1.7.4",
  "torchvision>=0.21.0,<0.22.0",
  "tqdm==4.67.1",
  "transformers==4.46.3",
  "voluptuous==0.15.2",

]

[project.optional-dependencies]
cu126 = [
    "torch==2.6.0+cu126",
    "torchvision==0.21.0+cu126"
]

[tool.uv]

[dependency-groups]
dev = ["ascii-magic==2.3.0", "matplotlib==3.10.0", "prompt-toolkit==3.0.51"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["takenoko"]

[tool.uv.sources]
torch = { index = "pytorch-cu126" }
torchvision = { index = "pytorch-cu126" }

[[tool.uv.index]]
name = "pytorch-cu126"
url = "https://download.pytorch.org/whl/cu126"
explicit = true

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = [
  "E", # pycodestyle errors
  "W", # pycodestyle warnings
  "F", # pyflakes
  "I", # isort
  "B", # flake8-bugbear
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "auto"
